+++ plugins/modules/netbox_ip_address.py
@@ -131,6 +131,12 @@ options:
               - The virtual machine the interface is attached to.
             type: str
             required: False
+      comments:
+        description:
+          - Comments that may include additional information in regards to the IP Address
+        required: false
+        type: str
+        version_added: "3.10.0"
       tags:
         description:
           - Any tags that the IP address may need to be associated with
@@ -163,35 +169,39 @@ EXAMPLES = r"""
 
   tasks:
     - name: Create IP address within NetBox with only required information
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
           address: 192.168.1.10
         state: present
+
     - name: Force to create (even if it already exists) the IP
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
           address: 192.168.1.10
         state: new
+
     - name: Get a new available IP inside 192.168.1.0/24
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
           prefix: 192.168.1.0/24
         state: new
+
     - name: Delete IP address within netbox
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
           address: 192.168.1.10
         state: absent
+
     - name: Create IP address with several specified options
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -204,8 +214,9 @@ EXAMPLES = r"""
           tags:
             - Schnozzberry
         state: present
+
     - name: Create IP address and assign a nat_inside IP
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -217,8 +228,9 @@ EXAMPLES = r"""
           interface:
             name: GigabitEthernet1
             device: test100
+
     - name: Ensure that an IP inside 192.168.1.0/24 is attached to GigabitEthernet1
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -228,8 +240,9 @@ EXAMPLES = r"""
             name: GigabitEthernet1
             device: test100
         state: present
+
     - name: Attach a new available IP of 192.168.1.0/24 to GigabitEthernet1
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -239,8 +252,9 @@ EXAMPLES = r"""
             name: GigabitEthernet1
             device: test100
         state: new
+
     - name: Attach a new available IP of 192.168.1.0/24 to GigabitEthernet1 (NetBox 2.9+)
-      netbox_ip_address:
+      netbox.netbox.netbox_ip_address:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -328,6 +342,7 @@ def main():
                             virtual_machine=dict(required=False, type="str"),
                         ),
                     ),
+                    comments=dict(required=False, type="str"),
                     tags=dict(required=False, type="list", elements="raw"),
                     custom_fields=dict(required=False, type="dict"),
                 ),