+++ plugins/modules/netbox_l2vpn.py
@@ -61,6 +61,12 @@ options:
           - The description of the L2VPN
         required: false
         type: str
+      comments:
+        description:
+          - Comments that may include additional information in regards to the L2VPN
+        required: false
+        type: str
+        version_added: "3.10.0"
       tenant:
         description:
           - The tenant that the L2VPN will be assigned to
@@ -88,7 +94,7 @@ EXAMPLES = r"""
 
   tasks:
     - name: Create L2VPN within NetBox with only required information
-      netbox_l2vpn:
+      netbox.netbox.netbox_l2vpn:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -97,7 +103,7 @@ EXAMPLES = r"""
         state: present
 
     - name: Delete L2VPN within netbox
-      netbox_vlan:
+      netbox.netbox.netbox_l2vpn:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -106,7 +112,7 @@ EXAMPLES = r"""
         state: absent
 
     - name: Create L2VPN with all required information
-      netbox_vlan:
+      netbox.netbox.netbox_l2vpn:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -163,6 +169,7 @@ def main():
                     import_targets=dict(required=False, type="list", elements="raw"),
                     export_targets=dict(required=False, type="list", elements="raw"),
                     description=dict(required=False, type="str"),
+                    comments=dict(required=False, type="str"),
                     tenant=dict(required=False, type="raw"),
                     tags=dict(required=False, type="list", elements="raw"),
                     custom_fields=dict(required=False, type="dict"),