+++ plugins/modules/netbox_prefix.py
@@ -99,6 +99,12 @@ options:
           - The description of the prefix
         required: false
         type: str
+      comments:
+        description:
+          - Comments that may include additional information in regards to the prefix
+        required: false
+        type: str
+        version_added: "3.10.0"
       tags:
         description:
           - Any tags that the prefix may need to be associated with
@@ -129,7 +135,7 @@ EXAMPLES = r"""
 
   tasks:
     - name: Create prefix within NetBox with only required information
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -137,7 +143,7 @@ EXAMPLES = r"""
         state: present
 
     - name: Delete prefix within netbox
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -145,7 +151,7 @@ EXAMPLES = r"""
         state: absent
 
     - name: Create prefix with several specified options
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -168,7 +174,7 @@ EXAMPLES = r"""
         state: present
 
     - name: Get a new /24 inside 10.156.0.0/19 within NetBox - Parent doesn't exist
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -178,7 +184,7 @@ EXAMPLES = r"""
         first_available: yes
 
     - name: Create prefix within NetBox with only required information
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -186,7 +192,7 @@ EXAMPLES = r"""
         state: present
 
     - name: Get a new /24 inside 10.156.0.0/19 within NetBox
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -196,7 +202,7 @@ EXAMPLES = r"""
         first_available: yes
 
     - name: Get a new /24 inside 10.157.0.0/19 within NetBox with additional values
-      netbox_prefix:
+      netbox.netbox.netbox_prefix:
         netbox_url: http://netbox.local
         netbox_token: thisIsMyToken
         data:
@@ -255,6 +261,7 @@ def main():
                     is_pool=dict(required=False, type="bool"),
                     mark_utilized=dict(required=False, type="bool"),
                     description=dict(required=False, type="str"),
+                    comments=dict(required=False, type="str"),
                     tags=dict(required=False, type="list", elements="raw"),
                     custom_fields=dict(required=False, type="dict"),
                 ),