+++ plugins/modules/netbox_route_target.py
@@ -44,6 +44,12 @@ options:
           - Tag description
         required: false
         type: str
+      comments:
+        description:
+          - Comments that may include additional information in regards to the route target
+        required: false
+        type: str
+        version_added: "3.10.0"
       tags:
         description:
           - Any tags that the device may need to be associated with
@@ -139,6 +145,7 @@ def main():
                     name=dict(required=True, type="str"),
                     tenant=dict(required=False, type="raw"),
                     description=dict(required=False, type="str"),
+                    comments=dict(required=False, type="str"),
                     tags=dict(required=False, type="list", elements="raw"),
                     custom_fields=dict(required=False, type="dict"),
                 ),