+++ plugins/modules/netbox_service.py
@@ -68,6 +68,12 @@ options:
           - Service description
         required: false
         type: str
+      comments:
+        description:
+          - Comments that may include additional information in regards to the service
+        required: false
+        type: str
+        version_added: "3.10.0"
       tags:
         description:
           - What tags to add/update
@@ -85,14 +91,14 @@ options:
 EXAMPLES = r"""
 - name: "Create netbox service"
   connection: local
-  hosts: all
+  hosts: localhost
   gather_facts: False
 
   tasks:
     - name: Create service
-      netbox_service:
-        netbox_url: netbox_url
-        netbox_token: netbox_token
+      netbox.netbox.netbox_service:
+        netbox_url: http://netbox.local
+        netbox_token: thisIsMyToken
         data:
           device: Test666
           name: node-exporter
@@ -104,16 +110,10 @@ EXAMPLES = r"""
             - prometheus
         state: present
 
-- name: "Delete netbox service"
-  connection: local
-  hosts: all
-  gather_facts: False
-
-  tasks:
     - name: Delete service
-      netbox_service:
-        netbox_url: netbox_url
-        netbox_token: netbox_token
+      netbox.netbox.netbox_service:
+        netbox_url: http://netbox.local
+        netbox_token: thisIsMyToken
         data:
           device: Test666
           name: node-exporter
@@ -152,6 +152,7 @@ def main():
                     protocol=dict(required=True, type="raw"),
                     ipaddresses=dict(required=False, type="raw"),
                     description=dict(required=False, type="str"),
+                    comments=dict(required=False, type="str"),
                     tags=dict(required=False, type="list", elements="raw"),
                     custom_fields=dict(required=False, type="dict"),
                 ),